requirements.txt
Flask==2.3.3
requests==2.31.0
google-generativeai==0.1.0rc1
groq==0.4.2
python-dotenv==1.0.0
Pillow==10.0.1
openai==1.3.7
Werkzeug==2.3.7
pypdf==3.17.1
pdf2image==1.16.3
python-docx==0.8.11
uuid==1.30 